requirements.txt
rdflib
pandas
numpy
matplotlib
pyyaml
requests
buildingmotif
typing
notebook
ipykernel
pipenv
